src/requirements.txt
Django==2.1.1
djangorestframework
django-sslserver
django_crontab
django-choices
python-lirc
py_irsend
pyHS100